+++ src/Transformer/FrameTransformer.php
@@ -31,7 +31,10 @@ class FrameTransformer extends AbstractTransformer
 
         if ($keys[0] == 'index') {
             $this->driver->switchTo()->frame($this->step['frame']['index']);
-        } else {
+        } elseif($keys[0] == 'default') {
+            $this->driver->switchTo()->defaultContent();
+        }
+        else {
             $element = $this->driver->findElement(WebLocator::get($keys[0], $values[0]));
             $this->driver->switchTo()->frame($element);
         }